Dat feature - api course

+++ controllers/course.controller.js
@@ -1,4 +1,6 @@
 const courseService = require("../services/course.service");
+const userService = require("../services/user.service");
+const courseJoinService = require("../services/course-join.service");
 
 const defaultRes = {
   result: 0,
@@ -22,7 +24,8 @@ module.exports = {
   async add(req, res, next) {
     const course = await courseService.addCourse(
       req.query.nameclass,
-      req.user.user_id
+      req.user.user_id,
+      req.query?.codeclass
     );
     res.json({
       ...defaultRes,
@@ -34,7 +37,14 @@ module.exports = {
 
   async delete(req, res, next) {
     const courseId = req.query.idclass;
-    await courseService.deleteCourseById(courseId);
+    const { user_id } = req.user;
+    const course = await courseService.getCourseById(courseId);
+
+    if (course && course.course_hostid === user_id) {
+      await courseJoinService.deleteCourseById(courseId);
+      await courseService.deleteCourseById(courseId);
+    }
+
     res.json({ result: {} });
   },
 
@@ -45,6 +55,40 @@ module.exports = {
       ...defaultRes,
       content: {
         course,
+      },
+    });
+  },
+
+  async updateInfo(req, res, next) {
+    const data = ({
+      course_name,
+      course_des,
+      course_topic,
+      course_code,
+      course_id,
+    } = req.body);
+    const { user_username } = req.user;
+
+    const course = await courseService.getCourseById(course_id);
+    const user = await userService.findUserByUsername2(user_username);
+
+    if (course && user && course_name) {
+      if (course.course_hostid === user.user_id) {
+        await courseService.updateInfo(data);
+
+        res.json({
+          ...defaultRes,
+          content: {
+            update: true,
+          },
+        });
+      }
+    }
+
+    res.json({
+      ...defaultRes,
+      content: {
+        update: false,
       },
     });
   },
